Refactor Accounts API
This commit refactors the accounts API mainly to allow updating the
manual_account_provider field. Prior to this change, the business
logic was such that once the account was created, the
account_provider could not be changed. Now that we have Plaid
linked accounts, the business logic needs to change to allow
updating an existing account with a manual_account_provider.

The main reason for this is because Plaid linked accounts will have
the opportunity to have a manual_account_provider added on. When
this happens, users will be able to manually upload statements that
Plaid was unable to fetch. For example, Plaid is only able to pull
the last 3 months of transactions from Capital One. To add
transactions older than 3 months ago, the user can enable
manual_account_provider on the account, and this will allow
statements to be parsed.

We also rename the account_provider attribute to
manual_account_provider to be more specific that it's related to
manual account parsing.

And finally, we add two new attributes to the AccountSerializer:
manualAccountEnabled and plaidAccountEnabled. These are boolean
values to indicate whether an account has a manual parser enabled
or a plaid linked account respectively.

+++ app/controllers/accounts_controller.rb
@@ -15,41 +15,39 @@ class AccountsController < ApplicationController
   end
 
   # POST /accounts
+  # API for creating manual accounts only. Plaid accounts are created via Plaid
+  # Link which then get synced via Plaid::SyncAccountsJob
   def create
-    account = AccountServices::Create.new(account_create_params).call
+    account = AccountServices::Create.new(account_params).call
     render json: account, status: :created, location: account
   end
 
   # PATCH/PUT /accounts/1
   def update
     account = Account.find(params[:id])
-    raise UnprocessableEntityError, account.errors unless account.update(account_update_params)
+    updated_account = AccountServices::Update.new(account, account_params).call
 
-    render json: account
+    render json: updated_account
+  rescue InvalidParser => e
+    error = {
+      manual_account_provider: ["'#{e.message}' is not a valid value."]
+    }
+    raise UnprocessableEntityError, error
   end
 
   # DELETE /accounts/1
   def destroy
     account = Account.find(params[:id])
+    if account.plaid_account_id.present?
+      raise BadRequestError, { account: ['Plaid accounts cannot be deleted at this time'] }
+    end
+
     account.destroy!
   end
 
   private
 
-  # Only allow a list of trusted parameters through for create action.
-  def account_create_params
-    params.permit(:account_provider, :active, :user_id)
-  end
-
-  # Only allow a list of trusted parameters through for update action.
-  def account_update_params
-    if params[:account_provider].present?
-      error = {
-        account_provider: ['cannot be updated after account creation.']
-      }
-      raise UnprocessableEntityError, error
-    end
-
-    params.permit(:active, :user_id)
+  def account_params
+    params.permit(:active, :manual_account_provider, :user_id)
   end
 end
